22-12-9

+++ 06 - Elágazások (switch-case)/Feladat 04/main.py
@@ -1,7 +1,7 @@
 num1:int=None
 num2:int=None
 math:str=None
-result:int=None
+result:float=None
 
 print("Kérek egy számot")
 num1=int(input())
